requirements.txt

python-telegram-bot==20.3
deepface==0.0.83
opencv-python
